build: build:esm-browser 脚本 改为 mjs
